// packages/docusaurus-sylk-plugin/src/sylk/protos/sylk/SylkMessage/v2/SylkMessage.d.ts
export interface SylkMessage {
    extensionType: SylkExtensions;
    name: string;
    uri: string;
    extensions: {
        [key: string]: {
            [key: string]: any;
        };
    };
    fullName: string;
    type: string;
    description: string;
    kind: string;
    fields: SylkField[];
    tag: string;
    inlines: Any[];
}